% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{fps_add_empty_factors}
\alias{fps_add_empty_factors}
\title{FPS: Add empty factors}
\usage{
fps_add_empty_factors(.data, factor_col = "cat", factor, factors_list_full)
}
\arguments{
\item{.data}{A data frame containing the analysis data.}

\item{factor_col}{A string specifying the column name in \code{.data} containing
the factor categories. Defaults to \code{"cat"}.}

\item{factor}{A string specifying the name of the factor to check for missing
levels.}

\item{factors_list_full}{A named list of all expected factor levels, where
each name corresponds to a factor, and the value is a vector of levels.}
}
\value{
A data frame with rows added for any missing factor levels, filled
with \code{NA} values.
}
\description{
This function ensures that all expected factor levels, including
"All farms," are present in a dataset. If a factor level is missing, an
empty row (filled with \code{NA}) is added to the dataset to preserve the
dataset dimensions.
}
\details{
If a category from \code{factors_list_full} is missing in the specified
\code{factor_col}, this function adds an empty row for that category. A warning
message is displayed for each missing category that is added.
}
\examples{
#data <- data.frame(cat = c("A", "B"), value = c(10, 20))
#factors_list_full <- list(my_factor = c("A", "B", "C"))
#updated_data <- fps_add_empty_factors(data, factor_col = "cat", factor = "my_factor", factors_list_full = factors_list_full)

}
\author{
Tom Pearson
}
