% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{fps_analyse_by_factor}
\alias{fps_analyse_by_factor}
\title{FPS: Analysis: Analyse by factor}
\usage{
fps_analyse_by_factor(
  design,
  factor,
  variable,
  denominator = NULL,
  ratio = FALSE,
  factor_col = "cat"
)
}
\arguments{
\item{design}{A survey design object created using the \code{survey} or \code{srvyr}
packages.}

\item{factor}{A character string specifying the name of the grouping variable
in the dataset.}

\item{variable}{A character string specifying the name of the binary variable
to analyze.}

\item{denominator}{(Optional) A character string specifying the denominator
variable for ratio calculations. Required if \code{ratio = TRUE}.}

\item{ratio}{A logical value indicating whether to calculate ratios (\code{TRUE})
or proportions (\code{FALSE}). Defaults to \code{FALSE}.}

\item{factor_col}{A character string specifying the column name for the
factor in the output. Defaults to \code{"cat"}.}
}
\value{
A data frame containing the computed statistics, including:
\itemize{
\item Unweighted counts.
\item Weighted proportions or ratios.
\item Confidence intervals.
}
}
\description{
Computes means (as proportions) or ratios for the Farm Practices
Survey (FPS) data using a survey design object. This function is designed
for use with \code{fps_prepare_results}.
}
\details{
\itemize{
\item This function assumes the \code{variable} is binary (e.g. \code{1} for presence and \code{0} for absence).
\item Confidence intervals are computed at a 95\% confidence level by default.
}
}
\examples{
#result <-
#fps_analyse_by_factor(
#  design = fps_design, #Assuming `fps_design` is a survey design object
#  factor = "fps_gor",
#  variable = "Q1_1",
#  ratio = FALSE)

}
\author{
Tom Pearson (adapted to srvyr methods)
}
