% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{fps_analysis}
\alias{fps_analysis}
\title{FPS: Analysis: High-level analysis}
\usage{
fps_analysis(
  .data,
  factor_col = "cat",
  questions_list,
  standard_factors_list,
  adhoc_factors_list = NULL,
  adhoc_factors_levels_list = NULL,
  ratio = FALSE,
  results_fp,
  tables_fp
)
}
\arguments{
\item{.data}{A data frame containing the survey data to be analysed. The data
frame must include relevant columns for defining the survey design (e.g.
\code{cph_no}, \code{post_strat}, \code{num_pop}).}

\item{factor_col}{A character string specifying the column containing the
categorical variable(s) used for grouping in the analysis. Defaults to
\code{"cat"}. The specified column should exist in \code{.data}.}

\item{questions_list}{A named list where each element corresponds to a survey
question. Each element must be a character vector consisting of:
\itemize{
\item The question identifier prefixed with \code{"answered_"} (e.g. \code{"answered_Q1"}).
\item Response options associated with the question.
}}

\item{standard_factors_list}{A named list of standard factors used for
grouping analysis. Each element of the list should be a named vector or a
list defining factor levels.}

\item{adhoc_factors_list}{An optional named list of ad-hoc factors to be
included in the analysis. These factors are added dynamically to
\code{standard_factors_list}. Defaults to \code{NULL}.}

\item{adhoc_factors_levels_list}{An optional named list specifying the levels
for the ad-hoc factors. This must align with \code{adhoc_factors_list} to ensure
consistency. Defaults to \code{NULL}.}

\item{ratio}{Logical. If \code{TRUE}, the analysis calculates ratios for the
specified questions using a denominator derived from the variable names in
\code{questions_list}. If \code{FALSE}, means are calculated. Defaults to
\code{FALSE}.}

\item{results_fp}{A character string specifying the file path where
intermediate analysis results will be saved as XLSX files. The path must
exist, and the function appends the results filenames.}

\item{tables_fp}{A character string specifying the file path where the final
formatted tables will be saved as XLSX files. The path must exist, and the
function appends the tables filenames.}
}
\value{
A named list (\code{table_list}) containing the formatted results for each
question. Each element is a nested list of data frames, one for each factor
analysed for the respective question.
}
\description{
This function performs high-level analysis for the Farm
Practices Survey (FPS). It processes survey data, sets the survey design,
and prepares results using lower-level functions (\code{fps_process_factors},
\code{fps_prepare_results}, and \code{fps_analyse_by_factor}). The analysis outputs
include formatted results and tables, written to XLSX files for quality
assurance and subsequent use in the RAP process.
}
\details{
The function performs the following steps:
\itemize{
\item Validates input arguments and ensures correct structure for all inputs.
\item Iterates through the questions specified in \code{questions_list} to subset data for each question.
\item Dynamically processes factors, including standard and ad-hoc factors, using \code{fps_process_factors}.
\item Sets the survey design using \code{srvyr::as_survey_design}.
\item Computes analysis results and writes intermediate outputs using \code{fps_prepare_results}.
\item Formats results into tables compatible with further RAP processing and ensures missing factor levels are explicitly represented using \code{fps_add_empty_factors}.
\item Writes the final tables to XLSX files in the location specified by \code{tables_fp}.
\item Factors with missing data are assigned \code{NA} values in outputs.
}
}
\examples{
#fps_analysis(
#  .data = s1_data,
#  factor_col = "cat",
#  questions_list = s1_questions_list,
#  standard_factors_list = factors_list,
#  adhoc_factors_list = s1_adhoc_factors_list,
#  adhoc_factors_levels_list = s1_adhoc_factors_levels_list,
#  ratio = FALSE,
#  results_fp = "results/s1_results",
#  tables_fp = "tables/s1_tables")

}
\seealso{
\code{\link[fps_process_factors]{fps_process_factors}},
\code{\link[fps_prepare_results]{fps_prepare_results}},
\code{\link[fps_analyse_by_factor]{fps_analyse_by_factor}}
}
\author{
Tom Pearson
}
