% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{fps_calc_weights}
\alias{fps_calc_weights}
\title{FPS: Calculate weights}
\usage{
fps_calc_weights(
  .data,
  pop_df,
  pop_col = "npop",
  unique_ref_col = "cph_no",
  join_col = "post_strat"
)
}
\arguments{
\item{.data}{A data frame containing survey data. Must include the columns
specified by \code{unique_ref_col} and \code{join_col}.}

\item{pop_df}{A data frame containing population data by stratum. Must
include columns specified by \code{join_col} and \code{pop_col}.}

\item{pop_col}{A string specifying the column in \code{pop_df} containing
population sizes for each stratum. Default is \code{"npop"}.}

\item{unique_ref_col}{A string specifying the column in \code{.data} containing
unique identifiers for observations. Default is \code{"cph_no"}.}

\item{join_col}{A string specifying the column in \code{.data} and \code{pop_df} used
to join the datasets (strata column). Default is \code{"post_strat"}.}
}
\value{
A data frame with the population column (\code{num_pop}) added to the
input data, and strata (\code{join_col}) converted to factors. Weights are
calculated but not included in the output, as the \code{survey} package
automatically computes these when using \code{fpc}.
}
\description{
This function calculates survey weights for Farm Practices
Survey (FPS) data by determining the number of observations per stratum,
joining it to the population data, and calculating weights as the ratio of
the population size to the number of observations. The population column is
then joined back into the dataset, and the strata are converted to factors.
}
\examples{
#survey_data <- data.frame(
#  cph_no = 1:10,
#  post_strat = rep(c("A", "B"), each = 5))
#
#population_data <- data.frame(
#  post_strat = c("A", "B"),
#  npop = c(100, 200))

}
\author{
Tom Pearson
}
