% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{fps_chart}
\alias{fps_chart}
\title{FPS: Chart}
\usage{
fps_chart(.data, questions, years, yaxis, fill, stacked = TRUE)
}
\arguments{
\item{.data}{A named list of data frames, where each list element corresponds
to a survey question and contains its associated data. Each data frame must
include columns \code{year}, \code{value}, and any variables specified by \code{yaxis} and
\code{fill}.}

\item{questions}{A character vector of one or more question identifiers
(e.g., "Q1", "Q2") to extract and visualize data for. Each identifier must
match a name in \code{.data}.}

\item{years}{A numeric vector of survey years to include in the chart.}

\item{yaxis}{A character string specifying the variable for the y-axis (e.g.
\code{"year"} or \code{"response"}).}

\item{fill}{A character string specifying the variable to use for fill colors
(e.g., \code{"year"} or \code{"response"}).}

\item{stacked}{Logical. If \code{TRUE}, the chart will use a stacked bar format;
if \code{FALSE}, it will use a dodged bar format. Defaults to \code{TRUE}.}
}
\value{
A ggplot object representing the bar chart.
}
\description{
This function generates custom bar charts for Farm Practices
Survey (FPS) data using ggplot2. The charts mimic GOV.UK auto-barchart
formatting and are intended for QA purposes.
}
\details{
The function:
\itemize{
\item Extracts and combines data for the specified questions.
\item Filters data by the given years.
\item Formats data for GOV.UK-style barcharts with appropriate aesthetics and color schemes.
\item Dynamically adjusts graph parameters, such as text size and bar positioning, based on the input data.
}
}
\examples{
#fps_chart(
#  .data = intro_tbl_list,
#  questions = "Q1",
#  years = c(2023, 2024),
#  yaxis = "year",
#  fill = "response",
#  stacked = TRUE)

}
\author{
Tom Pearson
}
