% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{fps_delete_old_versions}
\alias{fps_delete_old_versions}
\title{FPS: Render: Delete old versions (of reports/charts)}
\usage{
fps_delete_old_versions(
  out_dir,
  section = NULL,
  delete = FALSE,
  delete_unit = "hours",
  delete_number = 12
)
}
\arguments{
\item{out_dir}{Character. The base directory containing the \code{report} and
\code{charts} subdirectories.}

\item{section}{Character (optional). The specific section or subdirectory
under \code{report} and \code{charts} to process. Defaults to \code{NULL}, meaning no
specific section is targeted.}

\item{delete}{Logical. If \code{TRUE}, old files are deleted. Defaults to \code{FALSE}.}

\item{delete_unit}{Character. The time unit for the deletion threshold (e.g.
"secs", "mins", "hours", "days", "weeks"). See \link[base]{difftime}.
Defaults to \code{"hours"}.}

\item{delete_number}{Numeric. The number of \code{delete_unit} units defining the
threshold for deletion. Defaults to \code{12}.}
}
\description{
This function deletes old versions of reports and charts in the
specified output directory if the \code{delete} parameter is set to \code{TRUE}. The
deletion is based on a time period defined by \code{delete_unit} and
\code{delete_number}. The function is designed to be used prior to rendering a
new report to ensure only the current reports/charts are being used
}
\examples{
#fps_delete_old_versions(out_dir = "output", section = "01_nutrient_management",
#                        delete = TRUE, delete_unit = "days", delete_number = 30)

}
\seealso{
\link[base]{difftime}
}
\author{
Tom Pearson
}
