% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{fps_dodged}
\alias{fps_dodged}
\title{FPS: Charts: Dodged auto-barchart for GOV.UK}
\usage{
fps_dodged(.data, questions, svy_years, pivot_from_response = FALSE)
}
\arguments{
\item{.data}{A \code{tbl_list} for a specific section containing survey data for
different questions.}

\item{questions}{A character vector of questions (e.g. "Q1", "Q2", etc.) for
which the data should be processed. The function supports multiple
questions but assumes the data is comparable across them.}

\item{svy_years}{A vector of survey years to include in the chart (e.g.
\code{c(2020, 2021, 2022)}).}

\item{pivot_from_response}{A logical value. If \code{TRUE}, pivots the data from
the "response" variable to the response values. If \code{FALSE} (default),
pivots the data from the "year" variable to create columns for each year.}
}
\value{
A table formatted as a knitr object displaying the proportion of
responses in a dodged format. The chart can be pivoted either by response
or by year, depending on the value of \code{pivot_from_response}.
}
\description{
This function processes survey data to generate a table for use
on GOV.UK as a dodged chart using the \code{{barchart}} tag. The output displays
the proportion of responses for specified survey years, either by response
or by year. It allows for the visualization of how responses evolve across
survey years and can pivot the chart either by response or by year.
}
\details{
\itemize{
\item The function works with survey data stored in \code{tbl_list} for each section. Each table should represent a survey question.
\item It processes multiple questions by combining the data for all specified questions into one dataset.
\item The function handles two types of pivoting:
\itemize{
\item Pivoting by "response" (\code{pivot_from_response = TRUE}): This will create one column per response category.
\item Pivoting by "year" (\code{pivot_from_response = FALSE}): This will create one column per survey year.
}
\item It filters data for the specified survey years and formats the responses as percentages.
}
}
\examples{
#fps_dodged(survey_data, questions = c("Q1", "Q2"), svy_years = c(2020, 2021), pivot_from_response = TRUE)

}
\author{
Tom Pearson
}
