% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{fps_format_factors}
\alias{fps_format_factors}
\title{FPS: Pre-processing: Format factors}
\usage{
fps_format_factors(
  .data,
  region_col = "fps_nuts_name",
  farm_size_col = "fps_slr_name",
  farm_type_col = "fps_robust",
  strat_col = "post_strat"
)
}
\arguments{
\item{.data}{A data frame containing the data to format.}

\item{region_col}{A character string specifying the column name for the
region. Default is \code{"fps_nuts_name"}.}

\item{farm_size_col}{A character string specifying the column name for the
farm size. Default is \code{"fps_slr_name"}.}

\item{farm_type_col}{A character string specifying the column name for the
farm type. Default is \code{"fps_robust"}.}

\item{strat_col}{A numeric or character string specifying the column name for
the stratification. Default is \code{"post_strat"}.}
}
\value{
A data frame with formatted factor variables and renamed columns.
}
\description{
This function formats the factor variables in a dataset based on
standard FPS conventions. It renames specific columns for region, farm
size, farm type, and stratification to their standard names, applies
specific transformations to values in these columns, and ensures consistent
formatting.
}
\examples{
#data <- data.frame(
#  fps_gor = c("North West", "South East (England)", "Yorkshire"),
#  fps_slr_name = c("small", "medium", "large"),
#  fps_robust = c("Pigspoultry", "Dairy", "Arable"),
#  post_strat = c("1", "2", "3"))
#formatted_data <- fps_format_factors(data)

}
\author{
Tom Pearson
}
