% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{fps_make_allfarms_table}
\alias{fps_make_allfarms_table}
\title{FPS: Dataset: Make "All farms" table}
\usage{
fps_make_allfarms_table(table_list, allfarms_lookup_list, ratio = FALSE)
}
\arguments{
\item{table_list}{A named list of data frames containing the tables to be
processed.}

\item{allfarms_lookup_list}{A named list where names are the names of the
resulting "All farms" table, and values are named vectors mapping
sub-questions to their descriptive names (response options).}

\item{ratio}{Logical, defaults to \code{FALSE}. If \code{TRUE}, columns are
referenced assuming the analysis method was "ratio", otherwise "mean" is
used.}
}
\value{
A modified list of tables, including aggregated rows for the "All
Farms" category based on the lookup list.
}
\description{
This function processes a list of tables and integrates them
with a corresponding lookup list to create "All Farms" summary tables. It
is primarily used for \% questions as in section 5 (e.g. Q35; Q39) where a
comparison of the overall results across related questions is required
}
\examples{
#table_list <- example_table_list
#allfarms_lookup_list <-
#list("Q35_allfarms" = c("Q35_1" = "Stored under cover",
#                        "Q35_2" = "Stored in the open on a concret (impereable) base: covered",
#                        "Q35_3" = "Stored in the open on a concret (impereable) base: uncovered",
#                        "Q35_4" = "Stored in the open on a field site (no construction base): covered",
#                        "Q35_5" = "Stored in the open on a field site (no construction base): uncovered",
#                        "Q35_6" = "Store in any other type of facility"))
#result <- fps_make_allfarms_table(table_list, allfarms_lookup_list, ratio = FALSE)

}
\author{
Tom Pearson
}
