% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{fps_prepare_results}
\alias{fps_prepare_results}
\title{FPS: Analysis: Prepare results}
\usage{
fps_prepare_results(
  design,
  factors,
  variables,
  denominator = NULL,
  ratio = FALSE,
  factor_col = "cat",
  excel_file_path
)
}
\arguments{
\item{design}{A survey design object created using the \code{survey} or \code{srvyr}
packages.}

\item{factors}{A character vector specifying the grouping variables (factors)
for the analysis.}

\item{variables}{A character vector specifying the binary variables to
analyze.}

\item{denominator}{(Optional) A character string specifying the denominator
variable for ratio calculations. Required if \code{ratio = TRUE}.}

\item{ratio}{A logical value indicating whether to calculate ratios (\code{TRUE})
or means (\code{FALSE}). Defaults to \code{FALSE}.}

\item{factor_col}{A character string specifying the column name for factors
in the output. Defaults to \code{"cat"}.}

\item{excel_file_path}{A character string specifying the file path where the
Excel workbook will be saved.}
}
\value{
Saves the analysis results into an Excel file at the specified
\code{excel_file_path}.
}
\description{
This function calculates means or ratios for survey data
in binary format (using \code{fps_analyse_by_factor}), formats the results, and
saves them to an Excel file. The function iterates over multiple factors
and variables, performing the analysis and saving the results into separate
sheets (1 for each factor) in the output file.
}
\details{
\itemize{
\item Means are calculated when \code{ratio = FALSE} (default). Ratios are calculated when \code{ratio = TRUE}, and a denominator variable must be provided.
\item Results are saved in separate sheets named after each factor in the \code{factors} parameter.
\item Binary variables in \code{variables} should contain \code{1} for presence and \code{0} for absence.
\item For variables with "answered" in their names, warnings from convergence issues are suppressed to avoid irrelevant messages.
}
}
\examples{
#fps_prepare_results(
#  design = fpsdesign,
#  factors = c("region", "farm_type"),
#  variables = c("use_cover_crops", "answered"),
#  denominator = "total_area",
#  ratio = TRUE,
#  factor_col = "group",
#  excel_file_path = "results.xlsx")

}
\author{
Tom Pearson
}
