% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{fps_process_factors}
\alias{fps_process_factors}
\title{FPS: Analysis: Process factors}
\usage{
fps_process_factors(
  .data,
  question,
  factors,
  factors_list,
  adhoc_factors_list,
  adhoc_factors_levels_list
)
}
\arguments{
\item{.data}{A data frame containing the FPS data.}

\item{question}{A character string specifying the question of which to
process the factors for.}

\item{factors}{A character vector of existing factor column names to include
in the analysis.}

\item{factors_list}{A named list of factors with their levels to filter and
order the factors.}

\item{adhoc_factors_list}{A named list where each question maps to additional
ad-hoc factor column names. Defaults to \code{NULL}.}

\item{adhoc_factors_levels_list}{A named list mapping ad-hoc factor column
names to their levels. Defaults to \code{NULL}.}
}
\value{
A list with four components:
\describe{
\item{data}{The processed data frame with factors formatted.}
\item{factors}{An updated character vector of all factors included in the analysis.}
\item{factors_list}{A filtered and ordered named list of factors with levels used in the analysis.}
\item{factors_list_full}{A complete list of factors (including ad-hoc factors) with their levels.}
}
}
\description{
This function processes factors in the FPS dataset, including
standard factors (i.e. region, farm type, and farm size) and ad-hoc factors
provided by the user depending on the question. Ad-hoc factors are added to
the dataset, their levels are mapped and ordered, and both standard and
ad-hoc factors are prepared for analysis.
}
\examples{
#processed <-
#fps_process_factors(
#  .data = df,
#  question = "Q1",
#  factors = c("fps_gor", "fps_slr_name"),
#  factors_list = factors_list,
#  adhoc_factors_list = adhoc_factors_list,
#  adhoc_factors_levels_list = adhoc_factors_levels_list)

}
\author{
Tom Pearson
}
