% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{fps_remove_no_column}
\alias{fps_remove_no_column}
\title{FPS: Remove "No" Column}
\usage{
fps_remove_no_column(table_list, questions, questions_to_combine_list = NULL)
}
\arguments{
\item{table_list}{A named list of lists, where each list element corresponds
to a question and each list element within each question corresponds to a
table of factor-based results.}

\item{questions}{A character vector specifying question names to process
which should match the names of at least one of the tables in \code{table_list}.}

\item{questions_to_combine_list}{An optional named list. The names represent
the parent question to which a related question will be combined to. The
value is a single character vector containing the related question to
combine to the parent question.}
}
\value{
A modified version of \code{table_list} with processed columns and
optionally combined questions.
}
\description{
This function processes a list of tables (\code{table_list}) by
removing columns corresponding to "No" responses from specified questions,
ensuring data consistency for further analysis. Optionally, data for
questions with "No" responses removed can be combined with similarly
formatted questions
}
\examples{
#table_list <- list(Q9_exclNA = ..., Q18_exclNA = ...)
#questions <- c("Q9_exclNA", "Q18_exclNA")
#questions_to_combine_list <- list("Q9_exclNA" = "Q13_exclNA")
#result <- fps_remove_no_column(table_list, questions, questions_to_combine_list)

}
\author{
Tom Pearson
}
