% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{fps_stacked}
\alias{fps_stacked}
\title{FPS: Charts: Stacked auto-barchart for GOV.UK}
\usage{
fps_stacked(.data, questions, svy_years)
}
\arguments{
\item{.data}{A \code{tbl_list} for a specific section containing survey data for
different questions.}

\item{questions}{A character vector of questions (e.g. "Q1", "Q2", etc.) for
which the data should be processed. The function supports multiple
questions but assumes the data is comparable across them.}

\item{svy_years}{A vector of survey years to include in the chart (e.g.
\code{c(2020, 2021, 2022)}).}
}
\value{
A table formatted as a knitr object showing the proportion of
different responses in a stacked format, for each year in \code{svy_years}. Each
response is shown as a percentage of the total, with a "100\%" total column
added.
}
\description{
This function processes survey data to create a table for use on
GOV.UK as a stacked percentage chart using the \verb{\{barchart stacked\}} tag.
The output shows the proportions of different responses over the specified
years. It supports multiple questions and allows for comparison across
different survey years.
}
\details{
\itemize{
\item The function works with survey data stored in \code{tbl_list} for each section. Each table should represent a survey question.
\item It filters the data to include only the specified survey years, formats the response proportions as percentages, and arranges the data in a stacked format.
\item The input questions should be valid and exist in the \code{.data} list. If a question is not collected in the given years, the function will notify the user.
\item The function will process data for multiple questions by combining their data into one, ensuring the columns are compatible.
}
}
\examples{
#fps_stacked(survey_data, questions = c("Q1", "Q2"), svy_years = c(2020, 2021, 2022))

}
\author{
Tom Pearson
}
