% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{fps_write_dataset}
\alias{fps_write_dataset}
\title{FPS: Write dataset (XLSX)}
\usage{
fps_write_dataset(
  out_dir = ".",
  workbook,
  delete = FALSE,
  delete_unit = "hours",
  delete_number = 12,
  date = tolower(format(Sys.Date(), "\%d\%b\%y"))
)
}
\arguments{
\item{out_dir}{Character. The directory where the dataset should be saved.
Defaults to the current directory (\code{"."}).}

\item{workbook}{A \code{Workbook} object created with the \code{openxlsx} package. This
is the dataset to save.}

\item{delete}{Logical. If \code{TRUE}, old copies of the dataset in the output
directory will be deleted. Defaults to \code{FALSE}.}

\item{delete_unit}{Character. The time unit for deleting old files (e.g.,
\code{"hours"}, \code{"days"}). Defaults to \code{"hours"}.}

\item{delete_number}{Numeric. The number of time units to determine file age
for deletion. Files older than this value will be deleted. Defaults to
\code{12}.}

\item{date}{Character. The date string used to name the output file. Defaults
to today's date in the format \code{"\%d\%b\%y"} (e.g., "14jan25").}
}
\description{
This function saves the given dataset (workbook) to a file in
the specified output directory. The dataset is only saved once per run
regardless of how many sections are being knit. Optionally, old copies of
the dataset can be deleted based on a specified time period.
}
\details{
\itemize{
\item The Excel file is saved with the name \verb{fps-ghg-dataset-\{date\}.xlsx}.
\item Old files with \code{.xlsx} extensions in the output directory are deleted if \code{delete} is \code{TRUE} and they meet the age criteria.
\item File overwriting is enabled for the saved dataset.
}
}
\examples{
#fps_write_dataset(workbook = my_workbook, delete = TRUE, delete_unit = "hours", delete_number = 24)

}
\seealso{
\link[openxlsx]{saveWorkbook}
}
\author{
Tom Pearson
}
