% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrative.R
\name{get_doc}
\alias{get_doc}
\title{FPS: Narrative: Print the direction of change (doc) associated with
two values generated from the analysis results}
\usage{
get_doc(
  .data,
  question,
  response_key,
  svy_years,
  past = FALSE,
  abbr = FALSE,
  get_pc = FALSE
)
}
\arguments{
\item{.data}{A list of data frames, each representing survey question data.}

\item{question}{A character or numeric value specifying the question to
process. Numeric input is automatically converted to "QX" format (e.g.,
"Q4").}

\item{response_key}{A character vector of keywords for the desired
response(s). Supports regex patterns for filtering responses.}

\item{svy_years}{A numeric or character vector of length 2 specifying the
survey years for comparison (e.g., \code{c(2024, 2023)}).}

\item{past}{Logical. If \code{TRUE}, returns past-tense narrative (e.g.
"increased").}

\item{abbr}{Logical. If \code{TRUE}, returns abbreviated narrative (e.g. "an
increase" instead of "an increase from").}

\item{get_pc}{Logical. If \code{TRUE}, returns the percentage change instead of
narrative.}
}
\value{
A character string describing the direction of change or the
percentage change if \code{get_pc = TRUE}.
}
\description{
This function determines the direction of change (increase,
decrease, or no change) for a specified question and response in Farm
Practices Survey data, when comparing values between two survey years.
}
\details{
\itemize{
\item The comparison is always done from an older date to a newer date.
\item Values are rounded using \link[janitor]{round_half_up} before comparison.
\item Handles multiple response keys by summing their values unless the question is irregular.
}
}
\examples{
#data_list <- list(Q4 = data.frame(year = c("2024", "2023"),
#                                  value = c(60, 50),
#                                  response = c("yes", "yes")))
#get_doc(data_list, question = 4, response_key = "yes",
#        svy_years = c(2024, 2023), past = FALSE, abbr = FALSE, get_pc = TRUE)

}
\seealso{
\link[janitor]{round_half_up}
}
\author{
Tom Pearson
}
