% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrative.R
\name{get_name}
\alias{get_name}
\title{FPS: Narrative: Print the name/category associated with a value based
on its ordered position in the analysis' results.}
\usage{
get_name(
  .data,
  questions,
  svy_year,
  ordinal,
  get_percent = FALSE,
  responses_to_exclude = NULL
)
}
\arguments{
\item{.data}{A list of data frames, each representing survey question data.}

\item{questions}{A character or numeric vector specifying the question(s) to
process. Numeric input is automatically converted to "QX" format (e.g.
\code{"Q4"}).}

\item{svy_year}{A character or numeric value specifying the survey year
(e.g. \code{2024}).}

\item{ordinal}{A numeric value specifying the ordinal position (e.g. 1 for
the most frequent response).}

\item{get_percent}{Logical. If \code{TRUE}, returns the percentage of the selected
response instead of the response itself.}

\item{responses_to_exclude}{A character vector of responses to exclude from
the result. Responses are matched using regular expressions.}
}
\value{
A character string representing either the response or the
percentage, depending on \code{get_percent}.
}
\description{
This function retrieves the response or percentage corresponding
to the specified ordinal position (e.g. the most common response) for a
question from a list of survey data, ordering by the proportion of holdings
in descending order. You can also exclude specific responses and request
the result as a percentage rather than its name.
}
\details{
\itemize{
\item The data are ordered by the proportion of holdings in descending order (pre-rounding).
\item The function allows the exclusion of certain responses by their keywords.
\item The \code{ordinal} parameter specifies which position to retrieve (e.g., the most common response or the 2nd most common).
\item If \code{get_percent} is \code{TRUE}, the percentage of the selected response is returned.
}
}
\examples{
#data_list <- list(Q4 = data.frame(year = c("2024", "2024", "2024"),
#                                  value = c(50, 30, 20),
#                                  response = c("yes", "no", "maybe")))
#get_name(data_list, questions = 4, svy_year = 2024, ordinal = 1)
#get_name(data_list, questions = 4, svy_year = 2024, ordinal = 2, get_percent = TRUE)

}
\seealso{
\link[janitor]{round_half_up}
}
\author{
Tom Pearson
}
