% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrative.R
\name{round_number}
\alias{round_number}
\title{Round Numbers with Optional Prefix, Suffix, and Separator}
\usage{
round_number(val, prefix = NULL, suffix = NULL, big.mark = "")
}
\arguments{
\item{val}{A numeric vector of values to be rounded.}

\item{prefix}{A character string to be added before the number (optional).}

\item{suffix}{A character string to be added after the number (optional).}

\item{big.mark}{A character string to separate thousands (optional, default
is no mark, "").}
}
\value{
A character vector with the values rounded to whole numbers and
formatted according to the specified options.
}
\description{
This function rounds numeric values to the nearest whole number
and formats them with optional prefix, suffix, and thousands separator. It
uses the \code{scales::label_number()} function for formatting.
}
\details{
\itemize{
\item The \code{val} input must be numeric.
\item \code{prefix}, \code{suffix}, and \code{big.mark} are optional formatting parameters to control how the numbers are displayed.
}
}
\examples{
#round_number(12345.67)  # Returns "12346"
#round_number(12345.67, prefix = "£")  # Returns "£12346"
#round_number(12345.67, suffix = " units")  # Returns "12346 units"
#round_number(12345.67, big.mark = ",")  # Returns "12,346"
#round_number(12345.67, prefix = "£", suffix = " units", big.mark = ",")  # Returns "£12,346 units"

}
\author{
Tom Pearson
}
