% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_responses.R
\name{fps_name_responses}
\alias{fps_name_responses}
\title{FPS: Name responses: Label the coded responses in the analysis results}
\usage{
fps_name_responses(
  table_list,
  analysis_questions,
  response_lookup_df,
  survey_year = as.numeric(format(Sys.Date(), "\%Y")),
  ratio = FALSE,
  category = "All farms",
  factor = 1,
  rename = FALSE
)
}
\arguments{
\item{table_list}{A list of data frames containing analysis outputs.}

\item{analysis_questions}{A list or vector of questions analysed (must match
table names in \code{table_list}).}

\item{response_lookup_df}{A data frame mapping response codes to
human-readable names, including a "question" column.}

\item{survey_year}{Numeric. Defaults to the current system year (e.g. 2025).
This will be appended to the output table so that data can be distinguished
when adding to the time series later on.}

\item{ratio}{Logical. If \code{TRUE}, use "ratio" instead of "mean" as the
aggregation method assumed to be used in the analysis. Default is \code{FALSE}.}

\item{category}{Character. The category to filter by (e.g. "All farms").
Default is \code{"All farms"}.}

\item{factor}{The factor or grouping variable to extract from \code{table_list}.
Default is \code{1} which returns the value associated with the 1st factor in
the table (values are the same across all factors of "All farms").}

\item{rename}{Logical or character. If \code{TRUE}, attempts to rename the tables
based on extracted response patterns. If a character string, uses it as the
new name. Default is \code{FALSE}.}
}
\value{
A list of processed data frames with human-readable response names.
}
\description{
Joins coded variable names from analysis outputs to
human-readable names manually specified in a lookup dataframe, optionally
renames the tables, and processes them for narrative automation later on.
}
\examples{
#result <- fps_name_responses(
#  table_list = s8_livestock_table_list,
#  analysis_questions = s8_livestock_questions_list,
#  response_lookup_df = s8_livestock_response_lookup_df,
#  survey_year = 2025,
#  ratio = TRUE,
#  category = "dairy",
#  factor = "fps_robust",
#  rename = TRUE)

}
\author{
Tom Pearson
}
