% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{fps_update_dataset}
\alias{fps_update_dataset}
\title{FPS: Dataset: Update dataset}
\usage{
fps_update_dataset(
  table_list,
  questions,
  standard_factors_list,
  workbook,
  sheet,
  special_qs,
  adhoc_factors_list = NULL,
  adhoc_factors_levels_list = NULL,
  ratio = FALSE
)
}
\arguments{
\item{table_list}{A named list of data tables. Each table contains the
results for a specific question and factor.}

\item{questions}{A named list of row number vectors. Each element's
name (e.g. \code{"Q1"}) should correspond to a question in \code{table_list}. Each row number value
corresponds to the first row of the table in the dataset template that each
question’s data (for each factor) should be inserted.}

\item{standard_factors_list}{A named list of factors and their corresponding
levels for standard questions.}

\item{workbook}{An \code{openxlsx} workbook object where the data will be written.}

\item{sheet}{A character string specifying the name of the worksheet where
the data will be written.}

\item{special_qs}{A character vector of question names that require special
validation handling.}

\item{adhoc_factors_list}{An optional named list of ad-hoc factors to be
included for specific questions.}

\item{adhoc_factors_levels_list}{An optional named list of factor levels
corresponding to \code{adhoc_factors_list}.}

\item{ratio}{Logical. If TRUE, calculates ratios; if FALSE, calculates means
(default is FALSE).}
}
\value{
None. Updates the workbook in memory. The workbook must be saved
separately using \link[openxlsx]{saveWorkbook}.
}
\description{
This function processes survey results into a workbook,
formatting data and performing validation checks before inserting data into
a specific worksheet at specific row positions in the workbook. Ad-hoc
factors can be added to customize the processing for specific questions.
}
\examples{
#fps_update_dataset(table_list, questions, standard_factors_list, workbook, "Sheet1", special_qs)

}
\seealso{
\link[openxlsx]{saveWorkbook}
}
\author{
Tom Pearson
}
